requirements.txt
pandas
scikit-learn
catboost
pyyaml
joblib
numpy
optuna
psutil
seaborn
kaggle
scipy